+++ terminal-emulator/src/main/rust/tests/block_element_render_test.rs
@@ -87,7 +87,7 @@ fn char_wc_width_test(ucs: u32) -> usize {
 /// Verify block elements are stored correctly in screen buffer
 #[test]
 fn test_block_elements_in_screen_buffer() {
-    let mut engine = TerminalEngine::new(10, 5, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(10 as i64, 5 as i64, 1000, 10, 20);
 
     // Write block elements using UTF-8
     let text = "\u{2580}\u{2580}\u{2580}\u{2580}\u{2580}\u{2580}\u{2580}\u{2580}\u{2580}\u{2580}";
@@ -104,7 +104,7 @@ fn test_block_elements_in_screen_buffer() {
 /// Verify full block fills correctly
 #[test]
 fn test_full_block_rendering() {
-    let mut engine = TerminalEngine::new(5, 3, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(5 as i64, 3 as i64, 1000, 10, 20);
 
     let full_block = "\u{2588}";
     // Fill 3 lines, no scrolling yet
@@ -122,7 +122,7 @@ fn test_full_block_rendering() {
 /// Verify half blocks (used by neofetch, cava, etc.)
 #[test]
 fn test_half_block_pattern() {
-    let mut engine = TerminalEngine::new(10, 3, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(10 as i64, 3 as i64, 1000, 10, 20);
 
     let lower = "\u{2584}";
     engine.process_bytes(lower.repeat(10).as_bytes());
@@ -136,7 +136,7 @@ fn test_half_block_pattern() {
 /// Verify shade characters ( ░▒▓ )
 #[test]
 fn test_shade_characters() {
-    let mut engine = TerminalEngine::new(10, 3, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(10 as i64, 3 as i64, 1000, 10, 20);
 
     engine.process_bytes(b"\xe2\x96\x91\xe2\x96\x92\xe2\x96\x93\xe2\x96\x91\xe2\x96\x92\xe2\x96\x93\xe2\x96\x91\xe2\x96\x92\xe2\x96\x93");
 
@@ -149,7 +149,7 @@ fn test_shade_characters() {
 /// Verify box drawing forms a proper box
 #[test]
 fn test_box_drawing_form_box() {
-    let mut engine = TerminalEngine::new(6, 5, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(6 as i64, 5 as i64, 1000, 10, 20);
 
     // Use \r\n for each line to advance cursor (5 rows to avoid scrolling)
     engine.process_bytes(b"\xe2\x94\x8c\xe2\x94\x80\xe2\x94\x80\xe2\x94\x80\xe2\x94\x80\xe2\x94\x90\r\n");
@@ -218,7 +218,7 @@ fn test_block_element_font_selection() {
 /// Verify block elements survive scroll operations
 #[test]
 fn test_block_elements_survive_scroll() {
-    let mut engine = TerminalEngine::new(10, 5, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(10 as i64, 5 as i64, 1000, 10, 20);
 
     let full = "\u{2588}".repeat(10);
     let upper = "\u{2580}".repeat(10);
@@ -250,7 +250,7 @@ fn test_block_elements_survive_scroll() {
 /// Verify that a common neofetch-style output renders correctly
 #[test]
 fn test_neofetch_style_output() {
-    let mut engine = TerminalEngine::new(20, 10, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(20 as i64, 10 as i64, 1000, 10, 20);
 
     // Simulate a neofetch-like output using half blocks for color blocks
     engine.process_bytes(b"\x1b[31m\xe2\x96\x88\xe2\x96\x88\xe2\x96\x88\xe2\x96\x88\x1b[0m\x1b[32m\xe2\x96\x88\xe2\x96\x88\xe2\x96\x88\xe2\x96\x88\x1b[0m\x1b[34m\xe2\x96\x88\xe2\x96\x88\xe2\x96\x88\xe2\x96\x88\x1b[0m\r\n");
