+++ terminal-emulator/src/main/rust/tests/fix_verification.rs
@@ -24,7 +24,7 @@ fn get_row_text(engine: &TerminalEngine, row: i32) -> String {
 /// 验证 CJK 宽字符后的占位符 \0 不被计入空间使用
 #[test]
 fn test_get_space_used_ignores_null_placeholder() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入一个 CJK 宽字符 "你" (占 2 列，第二列是 \0 占位符)
     // UTF-8 bytes for "你": 0xE4 0xBD 0xA0
@@ -33,7 +33,7 @@ fn test_get_space_used_ignores_null_placeholder() {
 
     // 获取第一行
     let screen = &engine.state.main_screen;
-    let row = screen.get_row(0);
+    let row = screen.get_row(0 as i64);
     
     // 验证第一个字符是 "你"
     assert_eq!(row.text[0], '你', "First char should be '你'");
@@ -51,14 +51,14 @@ fn test_get_space_used_ignores_null_placeholder() {
 /// 验证多个 CJK 字符的占位符处理
 #[test]
 fn test_get_space_used_multiple_cjk() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入 "你好" (两个 CJK 宽字符)
     // UTF-8: 你 = 0xE4 0xBD 0xA0, 好 = 0xE5 0xA5 0xBD
     engine.process_bytes(b"\xe4\xbd\xa0\xe5\xa5\xbd");
 
     let screen = &engine.state.main_screen;
-    let row = screen.get_row(0);
+    let row = screen.get_row(0 as i64);
     
     // 验证字符布局
     assert_eq!(row.text[0], '你');
@@ -77,14 +77,14 @@ fn test_get_space_used_multiple_cjk() {
 /// 验证混合 ASCII 和 CJK 的空间计算
 #[test]
 fn test_get_space_used_mixed_ascii_cjk() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入 "Hi 你" (2 ASCII + 1 space + 1 CJK)
     // UTF-8: 你 = 0xE4 0xBD 0xA0
     engine.process_bytes(b"Hi \xe4\xbd\xa0");
 
     let screen = &engine.state.main_screen;
-    let row = screen.get_row(0);
+    let row = screen.get_row(0 as i64);
     
     // 布局应该是：'H', 'i', ' ', '你', '\0'
     assert_eq!(row.text[0], 'H');
@@ -108,17 +108,17 @@ fn test_get_space_used_mixed_ascii_cjk() {
 /// 验证 clear_all() 清空整行
 #[test]
 fn test_clear_all() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 先写入一些内容
     engine.process_bytes(b"Hello World");
     
     // 获取当前行并清空
     let screen = &mut engine.state.main_screen;
-    screen.get_row_mut(0).clear_all(0);
+    screen.get_row_mut(0 as i64).clear_all(0);
     
     // 验证整行被清空
-    let row = screen.get_row(0);
+    let row = screen.get_row(0 as i64);
     for i in 0..80 {
         assert_eq!(row.text[i], ' ', "Position {} should be space after clear_all", i);
         assert_eq!(row.styles[i], 0, "Style at {} should be 0 after clear_all", i);
@@ -130,21 +130,21 @@ fn test_clear_all() {
 /// 验证 clear_all() 不重置 line_wrap (对齐 Java)
 #[test]
 fn test_clear_all_preserves_line_wrap() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入超过 80 列的内容强制换行
     let long_text = "A".repeat(100);
     engine.process_bytes(long_text.as_bytes());
     
     // 第一行应该有 line_wrap = true
-    let row_before = engine.state.main_screen.get_row(0);
+    let row_before = engine.state.main_screen.get_row(0 as i64);
     let had_line_wrap = row_before.line_wrap;
     
     // 清空第一行
-    engine.state.main_screen.get_row_mut(0).clear_all(0);
+    engine.state.main_screen.get_row_mut(0 as i64).clear_all(0);
     
     // line_wrap 应该保持不变 (Java 行为)
-    let row_after = engine.state.main_screen.get_row(0);
+    let row_after = engine.state.main_screen.get_row(0 as i64);
     assert_eq!(row_after.line_wrap, had_line_wrap, "line_wrap should be preserved after clear_all");
     
     println!("✅ clear_all() preserves line_wrap flag");
@@ -157,7 +157,7 @@ fn test_clear_all_preserves_line_wrap() {
 /// 验证缩小屏幕时的内容重排
 #[test]
 fn test_resize_shrink_reflow() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入一行 80 字符
     let line = "A".repeat(80);
@@ -166,7 +166,7 @@ fn test_resize_shrink_reflow() {
     engine.process_bytes(b"Second line");
     
     // 缩小到 40 列
-    engine.state.resize(40, 24);
+    engine.state.resize(40 as i64, 24 as i64);
     
     // 验证内容被正确重排
     // 第一行应该是 40 个 'A'
@@ -183,7 +183,7 @@ fn test_resize_shrink_reflow() {
 /// 验证放大屏幕时的内容重排
 #[test]
 fn test_resize_expand_reflow() {
-    let mut engine = TerminalEngine::new(40, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(40 as i64, 24 as i64, 100, 10, 20);
 
     // 写入两行 40 字符
     let a_line = "A".repeat(40);
@@ -193,7 +193,7 @@ fn test_resize_expand_reflow() {
     engine.process_bytes(b_line.as_bytes());
 
     // 放大到 80 列
-    engine.state.resize(80, 24);
+    engine.state.resize(80 as i64, 24 as i64);
 
     // 验证内容被合并到一行
     let row0 = get_row_text(&engine, 0);
@@ -211,13 +211,13 @@ fn test_resize_expand_reflow() {
 /// 验证 resize 时光标位置追踪
 #[test]
 fn test_resize_cursor_tracking() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入内容
     engine.process_bytes(b"Hello World");
 
     // 缩小屏幕
-    engine.state.resize(40, 24);
+    engine.state.resize(40 as i64, 24 as i64);
 
     // 验证光标位置合理 (应该在屏幕内)
     assert!(
@@ -235,7 +235,7 @@ fn test_resize_cursor_tracking() {
 /// 验证 skipped_blank_lines 逻辑 (Java 对齐)
 #[test]
 fn test_resize_skipped_blank_lines() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入内容后跟空行
     engine.process_bytes(b"Line 1\r\n");
@@ -243,7 +243,7 @@ fn test_resize_skipped_blank_lines() {
     engine.process_bytes(b"Line 3\r\n");
     
     // 缩小屏幕
-    engine.state.resize(40, 24);
+    engine.state.resize(40 as i64, 24 as i64);
     
     // 验证内容存在
     let text = engine.state.main_screen.get_transcript_text();
@@ -260,13 +260,13 @@ fn test_resize_skipped_blank_lines() {
 /// 验证 CJK 字符在行尾的换行
 #[test]
 fn test_cjk_wrap_at_line_end() {
-    let mut engine = TerminalEngine::new(5, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(5 as i64, 24 as i64, 100, 10, 20);
 
     // 屏幕只有 5 列，写入 "AB 你" (2 + 2 = 4 列，应该能放下)
     // UTF-8: 你 = 0xE4 0xBD 0xA0
     engine.process_bytes(b"AB \xe4\xbd\xa0");
     
-    let row = engine.state.main_screen.get_row(0);
+    let row = engine.state.main_screen.get_row(0 as i64);
     
     // 验证字符位置
     assert_eq!(row.text[0], 'A');
@@ -281,21 +281,21 @@ fn test_cjk_wrap_at_line_end() {
 /// 验证 CJK 字符跨行换行
 #[test]
 fn test_cjk_wrap_across_lines() {
-    let mut engine = TerminalEngine::new(5, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(5 as i64, 24 as i64, 100, 10, 20);
 
     // 屏幕只有 5 列，写入 "AB 你好" (2 + 2 + 2 = 6 列，"好" 应该换行)
     // UTF-8: 你 = 0xE4 0xBD 0xA0, 好 = 0xE5 0xA5 0xBD
     engine.process_bytes(b"AB \xe4\xbd\xa0\xe5\xa5\xbd");
     
     // 第一行应该是 "AB 你\0"
-    let row0 = engine.state.main_screen.get_row(0);
+    let row0 = engine.state.main_screen.get_row(0 as i64);
     assert_eq!(row0.text[0], 'A');
     assert_eq!(row0.text[1], 'B');
     assert_eq!(row0.text[2], ' ');
     assert_eq!(row0.text[3], '你');
     
     // 第二行应该以 "好" 开始
-    let row1 = engine.state.main_screen.get_row(1);
+    let row1 = engine.state.main_screen.get_row(1 as i64);
     assert_eq!(row1.text[0], '好');
     
     println!("✅ CJK character wrapping across lines works correctly");
@@ -310,12 +310,12 @@ fn test_cjk_wrap_across_lines() {
 fn test_combining_char_space_calculation() {
     // 注意：Rust 侧目前不直接处理 combining chars
     // 这个测试验证基础行为
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入普通字符
     engine.process_bytes(b"e\xcc\x81"); // e + combining acute accent = é
     
-    let row = engine.state.main_screen.get_row(0);
+    let row = engine.state.main_screen.get_row(0 as i64);
     
     // 验证内容被写入
     let used = row.get_space_used();
@@ -331,7 +331,7 @@ fn test_combining_char_space_calculation() {
 /// 复杂场景：CJK + resize + 光标追踪
 #[test]
 fn test_complex_cjk_resize_cursor() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     // 写入混合内容
     // "Hello 世界！" - UTF-8: 世 = 0xE4 0xB8 96, 界 = 0xE7 0x95 8C, ！ = 0xEF 0xBC 0x81
@@ -345,9 +345,9 @@ fn test_complex_cjk_resize_cursor() {
     let _cursor_before_y = engine.state.cursor.y;
 
     // 多次 resize
-    engine.state.resize(40, 24);
-    engine.state.resize(60, 24);
-    engine.state.resize(80, 24);
+    engine.state.resize(40 as i64, 24 as i64);
+    engine.state.resize(60 as i64, 24 as i64);
+    engine.state.resize(80 as i64, 24 as i64);
 
     // 验证内容完整性 - 检查原始文本（\0 被替换为空格）
     let text = engine.state.main_screen.get_transcript_text();
@@ -369,12 +369,12 @@ fn test_complex_cjk_resize_cursor() {
 /// 边界情况：极窄屏幕 resize
 #[test]
 fn test_extreme_narrow_resize() {
-    let mut engine = TerminalEngine::new(80, 24, 100, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 100, 10, 20);
 
     engine.process_bytes(b"ABCDEFGHIJKLMNOPQRSTUVWXYZ");
     
     // 缩小到极窄 (10 列)
-    engine.state.resize(10, 24);
+    engine.state.resize(10 as i64, 24 as i64);
     
     // 验证内容被正确分割
     let row0 = get_row_text(&engine, 0);
